scalability framework: resolve circular dependency
